// mogaERP.Client/src/app/core/models/system-settings/customers.ts
export interface Customers {
    id: number
    name: string
    accountCode: string
    address: string
    phoneNumber: string
    taxNumber: string
    commercialRegistration: string
    paymentType: string
    creditLimit: number
    createdById: string
    createdBy: string
    createdOn: string
    updatedById: any
    updatedBy: any
    updatedOn: any
}  